Fix time class logic

+++ src/rpc/grpc/grpc_adapter.cpp
@@ -11,8 +11,7 @@ void GrpcAdapter::init(std::shared_ptr<BaseRpc> service)
     _service = service;
 }
 
-::grpc::Status GrpcAdapter::balance(
-    grpc::ServerContext*, const likelib::Address* request, likelib::Money* response)
+::grpc::Status GrpcAdapter::balance(grpc::ServerContext*, const likelib::Address* request, likelib::Money* response)
 {
     auto address = request->address().c_str();
 
@@ -32,7 +31,8 @@ void GrpcAdapter::init(std::shared_ptr<BaseRpc> service)
     auto from_address = request->from_address().address().c_str();
     auto to_address = request->to_address().address().c_str();
     auto amount = request->amount().money();
-    auto creation_time = std::stoul(request->creation_time().seconds_from_epoch());
+
+    auto creation_time = static_cast<std::uint_least32_t>(std::stoul(request->creation_time().seconds_from_epoch()));
 
     try {
         response->set_hash_string(
